docs(examples): add brutalist prev/next pager to example pages
A sheet-style pager at the bottom of every example detail page: prev and
next slugs with their № numbers, plus a center sheet counter, wrapping
at both ends. The examples record and its canonical sitemap ordering
move into a shared examplesIndex module so the index grid and pager
always agree on numbering; the index page and its load function now
consume the shared module.
